requirements.txt
cs[async,highlight]
pydantic
python-dotenv
python-keycloak==5.4.0
pandas
fastapi 
uvicorn